+++ packages/common/web/src/Utils/viewerUtils.tsx
@@ -1,15 +1,29 @@
+import React from 'react';
+import { InlineStyleMapperFunction, InlineStyleMapper } from '../types';
 import { safeJsonParse } from './jsonUtils';
 import {
   RICOS_TEXT_COLOR_TYPE,
   RICOS_TEXT_HIGHLIGHT_TYPE,
   RICOS_FONT_SIZE_TYPE,
+  TEXT_HIGHLIGHT_TYPE,
+  TEXT_COLOR_TYPE,
+  DraftContent,
 } from 'ricos-content';
 import { isHexColor } from './colorUtils';
+import { getBlocksFromContentState } from './innerRCEBlocksUtils';
 
 export const INLINE_STYLE_TYPES = {
   [RICOS_TEXT_COLOR_TYPE]: 'FG',
   [RICOS_TEXT_HIGHLIGHT_TYPE]: 'BG',
   [RICOS_FONT_SIZE_TYPE]: 'font-size',
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type PluginConfig = Record<string, any>;
+
+const RICOS_TO_DRAFT_TYPE = {
+  [RICOS_TEXT_COLOR_TYPE]: TEXT_COLOR_TYPE,
+  [RICOS_TEXT_HIGHLIGHT_TYPE]: TEXT_HIGHLIGHT_TYPE,
 };
 
 export type CustomInlineStyleType = keyof typeof INLINE_STYLE_TYPES;
@@ -46,4 +60,36 @@ export const defaultStyleFnMapper = {
     const parsedStyle = dynamicStyleParsers[RICOS_FONT_SIZE_TYPE](style);
     return parsedStyle ? { fontSize: parsedStyle } : undefined;
   },
+};
+
+export const getDynamicInlineStyleMapper = (
+  type: CustomInlineStyleType
+): InlineStyleMapperFunction<PluginConfig> => {
+  return (config: PluginConfig, raw: DraftContent) => {
+    const settings = config[RICOS_TO_DRAFT_TYPE[type] || ''] || {};
+    const styleParser = dynamicStyleParsers[type];
+    const styleSelectionPredicate = settings.styleSelectionPredicate
+      ? (style: string) => styleParser(style, settings.styleSelectionPredicate)
+      : (style: string) => !!styleParser(style);
+    const customStyleFn = settings.customStyleFn
+      ? (style: string) =>
+          settings.customStyleFn(styleParser(style, settings.styleSelectionPredicate))
+      : defaultStyleFnMapper[type];
+    const rawBlocks = getBlocksFromContentState(raw);
+    const mapper = rawBlocks.reduce<InlineStyleMapper>((map, block) => {
+      if (block.inlineStyleRanges) {
+        block.inlineStyleRanges
+          .filter(range => styleSelectionPredicate(range.style))
+          .forEach(range => {
+            map[range.style] = (children, { key }) => (
+              <span key={key} style={customStyleFn(range.style)}>
+                {children}
+              </span>
+            );
+          });
+      }
+      return map;
+    }, {});
+    return () => mapper;
+  };
 };