+++ packages/viewer/web/src/inlineStyleMapper.tsx
@@ -1,58 +1,10 @@
 import React from 'react';
 import {
-  defaultStyleFnMapper,
-  TEXT_HIGHLIGHT_TYPE,
-  TEXT_COLOR_TYPE,
   RICOS_FONT_SIZE_TYPE,
   RICOS_TEXT_COLOR_TYPE,
   RICOS_TEXT_HIGHLIGHT_TYPE,
-  getBlocksFromContentState,
-  InlineStyleMapper,
-  DraftContent,
-  InlineStyleMapperFunction,
-  CustomInlineStyleType,
-  dynamicStyleParsers,
+  getDynamicInlineStyleMapper,
 } from 'wix-rich-content-common';
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type PluginConfig = Record<string, any>;
-
-const RICOS_TO_DRAFT_TYPE = {
-  [RICOS_TEXT_COLOR_TYPE]: TEXT_COLOR_TYPE,
-  [RICOS_TEXT_HIGHLIGHT_TYPE]: TEXT_HIGHLIGHT_TYPE,
-};
-
-const getDynamicInlineStyleMapper = (
-  type: CustomInlineStyleType
-): InlineStyleMapperFunction<PluginConfig> => {
-  return (config: PluginConfig, raw: DraftContent) => {
-    const settings = config[RICOS_TO_DRAFT_TYPE[type] || ''] || {};
-    const styleParser = dynamicStyleParsers[type];
-    const styleSelectionPredicate = settings.styleSelectionPredicate
-      ? (style: string) => styleParser(style, settings.styleSelectionPredicate)
-      : (style: string) => !!styleParser(style);
-    const customStyleFn = settings.customStyleFn
-      ? (style: string) =>
-          settings.customStyleFn(styleParser(style, settings.styleSelectionPredicate))
-      : defaultStyleFnMapper[type];
-    const rawBlocks = getBlocksFromContentState(raw);
-    const mapper = rawBlocks.reduce<InlineStyleMapper>((map, block) => {
-      if (block.inlineStyleRanges) {
-        block.inlineStyleRanges
-          .filter(range => styleSelectionPredicate(range.style))
-          .forEach(range => {
-            map[range.style] = (children, { key }) => (
-              <span key={key} style={customStyleFn(range.style)}>
-                {children}
-              </span>
-            );
-          });
-      }
-      return map;
-    }, {});
-    return () => mapper;
-  };
-};
 
 const textColorInlineStyleMapper = getDynamicInlineStyleMapper(RICOS_TEXT_COLOR_TYPE);