+++ CHANGELOG.md
@@ -17,7 +17,10 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
 
+- `*`
+  - [#2919](https://github.com/wix/ricos/pull/2919) missing plugin buttons in toolbar
 
 </details>
 <hr/>
@@ -34,8 +37,9 @@
 ### :house: Internal
 
 - `ui-components`
+
   - [#2908](https://github.com/wix/ricos/pull/2908) refactor button component - removed classname prop
-  
+
   ### :bug: Bug Fix
 
 - `toolbars-new`