+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 ### :house: Internal
 
 - `ui-components`
+
   - [#2908](https://github.com/wix/ricos/pull/2908) refactor button component - removed classname prop
 
   ### :bug: Bug Fix